Resume_Scr/requirements.txt
streamlit
pdfplumber
python-docx
nltk
spacy
scikit-learn
pandas
numpy
openpyxl
plotly
textract
python-magic-bin
PyPDF2
phonenumbers
dateparser